requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
APScheduler==3.6.3
async-timeout==4.0.2
attrs==22.2.0
black==23.1.0
cachetools==4.2.2
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
frozenlist==1.3.3
gTTS==2.3.1
idna==3.4
multidict==6.0.4
mypy-extensions==1.0.0
openai==0.27.0
packaging==23.0
pathspec==0.11.0
platformdirs==3.0.0
python-telegram-bot==13.13
pytube==12.1.2
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
requests==2.28.2
six==1.16.0
tornado==6.2
tqdm==4.64.1
tzdata==2022.7
tzlocal==4.2
urllib3==1.26.14
yarl==1.8.2
youtube-search==2.1.2
